+++ components/push/storage/src/schema.sql
@@ -1,5 +1,6 @@
--- XXX: maybe push_sub?
-CREATE TABLE IF NOT EXISTS push_record (
+CREATE TABLE
+IF NOT EXISTS push_record
+(
     uaid               TEXT     NOT NULL,
     channel_id         TEXT     NOT NULL UNIQUE,
     endpoint           TEXT     NOT NULL UNIQUE,
@@ -10,4 +11,10 @@ CREATE TABLE IF NOT EXISTS push_record (
     native_id          TEXT,
     PRIMARY KEY (uaid, channel_id)
 );
-CREATE UNIQUE INDEX channel_id_idx ON push_record(channel_id);
+
+CREATE TABLE
+IF NOT EXISTS meta_data
+(
+    key                TEXT    PRIMARY KEY,
+    value                      NOT NULL
+) without ROWID;